add_inode: fix shift bug

+++ src/ntfs2btrfs.cpp
@@ -2171,7 +2171,7 @@ static void add_inode(root& r, uint64_t inode, uint64_t ntfs_inode, bool& is_dir
 
     if (compression_unit != 0) {
         string compdata;
-        uint64_t cus = 1 << compression_unit;
+        uint64_t cus = 1 << (uint64_t)compression_unit;
 
         compdata.resize((size_t)(cus * cluster_size));
